DriverKit update

+++ projects/init/devel/Sel4_Stubs/simple/simple.h
@@ -15,4 +15,4 @@ typedef struct
 
 
 typedef void*  seL4_CPtr;
-typedef long seL4_Word ;
+typedef long long seL4_Word ;
